Shared Item action menu.
